executive-order changes and register-to-vote added
executive-order command now requires the user to have the President Role. I have also added the register-to-vote command which will add the Voter role to the user if they do not have it, if the role is not present on the server it will be created. The voter role will be required to vote in the future.